Refactor breadcrumbs: move computeBreadcrumb to Breadcrumb component and add css.
