+++ src/components/Breadcrumbs.css
@@ -1,0 +1,14 @@
+
+/*.Breadcrumbs {
+	text-transform: uppercase;
+}*/
+
+.BreadcrumbsContainer{
+  height: 65px;
+  display: flex;
+  flex-wrap: nowrap;
+  align-items: center;
+  font-size: 1.2em;
+  overflow-x: hidden;
+  -ms-overflow-style: -ms-autohiding-scrollbar;
+}
